requirements.txt
#for BirdeNet-Analyser
tflite-runtime
librosa
bottle

#For this app
pandas
pyalsaaudio
numpy
scipy
influxdb-client
pyTelegramBotAPI
pydub
pyyaml

